requirements.txt
# Core dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
groq==0.27.0
loguru==0.7.2
PyPDF2==3.0.1
PyYAML==6.0.1
nltk==3.8.1
SQLAlchemy==2.0.28
requests==2.31.0

# Machine learning
scikit-learn==1.7.0
numpy==1.26.2

# Web scraping
lxml==5.1.0
html5lib==1.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings>=0.24.0

# PDF Processing
PyMuPDF==1.23.26

# Web Scraping and Automation
playwright==1.42.0

# Data Processing
pandas==2.1.4
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.0.1

# Pydantic
pydantic==2.6.3

# New dependencies
matplotlib==3.8.2
selenium==4.16.0
webdriver-manager==4.0.1
seaborn==0.13.0
python-dotenv==1.0.0
tqdm==4.66.1 